+++ api-server/api/models/component.go
@@ -76,7 +76,7 @@ func buildComponent(
 	}
 
 	if jobComponent, ok := radixComponent.(*radixv1.RadixDeployJobComponent); ok {
-		builder.WithSchedulerPort(&jobComponent.SchedulerPort)
+		builder.WithSchedulerPort(jobComponent.SchedulerPort)
 		if jobComponent.Payload != nil {
 			builder.WithScheduledJobPayloadPath(jobComponent.Payload.Path)
 		}