+++ pipeline-runner/steps/applyconfig/step_test.go
@@ -161,8 +161,8 @@ func (s *applyConfigTestSuite) Test_Deploy_BuildJobInDeployPipelineShouldFail() 
 	ra := utils.NewRadixApplicationBuilder().
 		WithAppName(appName).
 		WithJobComponents(
-			utils.NewApplicationJobComponentBuilder().WithSchedulerPort(9999).WithName("buildjob"),
-			utils.NewApplicationJobComponentBuilder().WithSchedulerPort(9999).WithName("deployjob").WithImage("any:latest"),
+			utils.NewApplicationJobComponentBuilder().WithSchedulerPort(new(int32(9999))).WithName("buildjob"),
+			utils.NewApplicationJobComponentBuilder().WithSchedulerPort(new(int32(9999))).WithName("deployjob").WithImage("any:latest"),
 		).
 		WithEnvironment("dev", "anybranch").
 		BuildRA()
@@ -224,28 +224,28 @@ func (s *applyConfigTestSuite) Test_BuildAndDeployComponentImages() {
 				),
 		).
 		WithJobComponents(
-			utils.NewApplicationJobComponentBuilder().WithPort("any", 8080).WithName("job-1").WithSourceFolder("./client/").WithDockerfileName("client.Dockerfile").WithSchedulerPort(9999).
+			utils.NewApplicationJobComponentBuilder().WithPort("any", 8080).WithName("job-1").WithSourceFolder("./client/").WithDockerfileName("client.Dockerfile").WithSchedulerPort(new(int32(9999))).
 				WithEnvironmentConfigs(
 					utils.NewJobComponentEnvironmentBuilder().WithEnvironment(envName1),
 					utils.NewJobComponentEnvironmentBuilder().WithEnvironment(envName2).WithSourceFolder("./client2/"),
 					utils.NewJobComponentEnvironmentBuilder().WithEnvironment(envName3).WithDockerfileName("client2.Dockerfile"),
 					utils.NewJobComponentEnvironmentBuilder().WithEnvironment(envName4).WithImage("some-image2:some-tag"),
 				),
-			utils.NewApplicationJobComponentBuilder().WithPort("any", 8080).WithName("job-2").WithDockerfileName("client.Dockerfile").WithSchedulerPort(9999).
+			utils.NewApplicationJobComponentBuilder().WithPort("any", 8080).WithName("job-2").WithDockerfileName("client.Dockerfile").WithSchedulerPort(new(int32(9999))).
 				WithEnvironmentConfigs(
 					utils.NewJobComponentEnvironmentBuilder().WithEnvironment(envName1),
 					utils.NewJobComponentEnvironmentBuilder().WithEnvironment(envName2).WithSourceFolder("./client2/"),
 					utils.NewJobComponentEnvironmentBuilder().WithEnvironment(envName3).WithDockerfileName("client2.Dockerfile"),
 					utils.NewJobComponentEnvironmentBuilder().WithEnvironment(envName4).WithImage("some-image3:some-tag"),
 				),
-			utils.NewApplicationJobComponentBuilder().WithPort("any", 8080).WithName("job-3").WithSourceFolder("./client/").WithSchedulerPort(9999).
+			utils.NewApplicationJobComponentBuilder().WithPort("any", 8080).WithName("job-3").WithSourceFolder("./client/").WithSchedulerPort(new(int32(9999))).
 				WithEnvironmentConfigs(
 					utils.NewJobComponentEnvironmentBuilder().WithEnvironment(envName1),
 					utils.NewJobComponentEnvironmentBuilder().WithEnvironment(envName2).WithSourceFolder("./client2/"),
 					utils.NewJobComponentEnvironmentBuilder().WithEnvironment(envName3).WithDockerfileName("client2.Dockerfile"),
 					utils.NewJobComponentEnvironmentBuilder().WithEnvironment(envName4).WithImage("some-image4:some-tag"),
 				),
-			utils.NewApplicationJobComponentBuilder().WithPort("any", 8080).WithName("job-4").WithImage("some-image1:some-tag").WithSchedulerPort(9999).
+			utils.NewApplicationJobComponentBuilder().WithPort("any", 8080).WithName("job-4").WithImage("some-image1:some-tag").WithSchedulerPort(new(int32(9999))).
 				WithEnvironmentConfigs(
 					utils.NewJobComponentEnvironmentBuilder().WithEnvironment(envName1),
 					utils.NewJobComponentEnvironmentBuilder().WithEnvironment(envName2).WithSourceFolder("./client2/"),
@@ -410,25 +410,25 @@ func (s *applyConfigTestSuite) Test_BuildAndDeployComponentImages_ExpectedRuntim
 				WithEnvironmentConfig(utils.NewComponentEnvironmentBuilder().WithEnvironment("otherenv").WithRuntime(&radixv1.Runtime{Architecture: radixv1.RuntimeArchitectureAmd64})),
 		).
 		WithJobComponents(
-			utils.NewApplicationJobComponentBuilder().WithName("job1-build").WithSchedulerPort(9999),
-			utils.NewApplicationJobComponentBuilder().WithName("job2-build").WithRuntime(&radixv1.Runtime{Architecture: ""}).WithSchedulerPort(9999),
-			utils.NewApplicationJobComponentBuilder().WithName("job3-build").WithRuntime(&radixv1.Runtime{Architecture: radixv1.RuntimeArchitectureAmd64}).WithSchedulerPort(9999),
-			utils.NewApplicationJobComponentBuilder().WithName("job4-build").WithRuntime(&radixv1.Runtime{Architecture: radixv1.RuntimeArchitectureArm64}).WithSchedulerPort(9999),
-			utils.NewApplicationJobComponentBuilder().WithName("job5-build").WithRuntime(&radixv1.Runtime{Architecture: radixv1.RuntimeArchitectureArm64}).WithSchedulerPort(9999).
+			utils.NewApplicationJobComponentBuilder().WithName("job1-build").WithSchedulerPort(new(int32(9999))),
+			utils.NewApplicationJobComponentBuilder().WithName("job2-build").WithRuntime(&radixv1.Runtime{Architecture: ""}).WithSchedulerPort(new(int32(9999))),
+			utils.NewApplicationJobComponentBuilder().WithName("job3-build").WithRuntime(&radixv1.Runtime{Architecture: radixv1.RuntimeArchitectureAmd64}).WithSchedulerPort(new(int32(9999))),
+			utils.NewApplicationJobComponentBuilder().WithName("job4-build").WithRuntime(&radixv1.Runtime{Architecture: radixv1.RuntimeArchitectureArm64}).WithSchedulerPort(new(int32(9999))),
+			utils.NewApplicationJobComponentBuilder().WithName("job5-build").WithRuntime(&radixv1.Runtime{Architecture: radixv1.RuntimeArchitectureArm64}).WithSchedulerPort(new(int32(9999))).
 				WithEnvironmentConfig(utils.NewJobComponentEnvironmentBuilder().WithEnvironment(envName).WithRuntime(&radixv1.Runtime{})),
-			utils.NewApplicationJobComponentBuilder().WithName("job6-build").WithRuntime(&radixv1.Runtime{Architecture: radixv1.RuntimeArchitectureArm64}).WithSchedulerPort(9999).
+			utils.NewApplicationJobComponentBuilder().WithName("job6-build").WithRuntime(&radixv1.Runtime{Architecture: radixv1.RuntimeArchitectureArm64}).WithSchedulerPort(new(int32(9999))).
 				WithEnvironmentConfig(utils.NewJobComponentEnvironmentBuilder().WithEnvironment(envName).WithRuntime(&radixv1.Runtime{Architecture: radixv1.RuntimeArchitectureAmd64})),
-			utils.NewApplicationJobComponentBuilder().WithName("job7-build").WithRuntime(&radixv1.Runtime{Architecture: radixv1.RuntimeArchitectureArm64}).WithSchedulerPort(9999).
+			utils.NewApplicationJobComponentBuilder().WithName("job7-build").WithRuntime(&radixv1.Runtime{Architecture: radixv1.RuntimeArchitectureArm64}).WithSchedulerPort(new(int32(9999))).
 				WithEnvironmentConfig(utils.NewJobComponentEnvironmentBuilder().WithEnvironment("otherenv").WithRuntime(&radixv1.Runtime{Architecture: radixv1.RuntimeArchitectureAmd64})),
-			utils.NewApplicationJobComponentBuilder().WithName("job1-deploy").WithImage("any").WithSchedulerPort(9999),
-			utils.NewApplicationJobComponentBuilder().WithName("job2-deploy").WithImage("any").WithRuntime(&radixv1.Runtime{Architecture: ""}).WithSchedulerPort(9999),
-			utils.NewApplicationJobComponentBuilder().WithName("job3-deploy").WithImage("any").WithRuntime(&radixv1.Runtime{Architecture: radixv1.RuntimeArchitectureAmd64}).WithSchedulerPort(9999),
-			utils.NewApplicationJobComponentBuilder().WithName("job4-deploy").WithImage("any").WithRuntime(&radixv1.Runtime{Architecture: radixv1.RuntimeArchitectureArm64}).WithSchedulerPort(9999),
-			utils.NewApplicationJobComponentBuilder().WithName("job5-deploy").WithImage("any").WithRuntime(&radixv1.Runtime{Architecture: radixv1.RuntimeArchitectureArm64}).WithSchedulerPort(9999).
+			utils.NewApplicationJobComponentBuilder().WithName("job1-deploy").WithImage("any").WithSchedulerPort(new(int32(9999))),
+			utils.NewApplicationJobComponentBuilder().WithName("job2-deploy").WithImage("any").WithRuntime(&radixv1.Runtime{Architecture: ""}).WithSchedulerPort(new(int32(9999))),
+			utils.NewApplicationJobComponentBuilder().WithName("job3-deploy").WithImage("any").WithRuntime(&radixv1.Runtime{Architecture: radixv1.RuntimeArchitectureAmd64}).WithSchedulerPort(new(int32(9999))),
+			utils.NewApplicationJobComponentBuilder().WithName("job4-deploy").WithImage("any").WithRuntime(&radixv1.Runtime{Architecture: radixv1.RuntimeArchitectureArm64}).WithSchedulerPort(new(int32(9999))),
+			utils.NewApplicationJobComponentBuilder().WithName("job5-deploy").WithImage("any").WithRuntime(&radixv1.Runtime{Architecture: radixv1.RuntimeArchitectureArm64}).WithSchedulerPort(new(int32(9999))).
 				WithEnvironmentConfig(utils.NewJobComponentEnvironmentBuilder().WithEnvironment(envName).WithRuntime(&radixv1.Runtime{})),
-			utils.NewApplicationJobComponentBuilder().WithName("job6-deploy").WithImage("any").WithRuntime(&radixv1.Runtime{Architecture: radixv1.RuntimeArchitectureArm64}).WithSchedulerPort(9999).
+			utils.NewApplicationJobComponentBuilder().WithName("job6-deploy").WithImage("any").WithRuntime(&radixv1.Runtime{Architecture: radixv1.RuntimeArchitectureArm64}).WithSchedulerPort(new(int32(9999))).
 				WithEnvironmentConfig(utils.NewJobComponentEnvironmentBuilder().WithEnvironment(envName).WithRuntime(&radixv1.Runtime{Architecture: radixv1.RuntimeArchitectureAmd64})),
-			utils.NewApplicationJobComponentBuilder().WithName("job7-deploy").WithImage("any").WithRuntime(&radixv1.Runtime{Architecture: radixv1.RuntimeArchitectureArm64}).WithSchedulerPort(9999).
+			utils.NewApplicationJobComponentBuilder().WithName("job7-deploy").WithImage("any").WithRuntime(&radixv1.Runtime{Architecture: radixv1.RuntimeArchitectureArm64}).WithSchedulerPort(new(int32(9999))).
 				WithEnvironmentConfig(utils.NewJobComponentEnvironmentBuilder().WithEnvironment("otherenv").WithRuntime(&radixv1.Runtime{Architecture: radixv1.RuntimeArchitectureAmd64})),
 		).
 		BuildRA()
@@ -581,19 +581,19 @@ func (s *applyConfigTestSuite) Test_BuildAndDeployComponentImages_IgnoreDisabled
 				WithEnvironmentConfig(utils.NewComponentEnvironmentBuilder().WithEnvironment(envName).WithEnabled(true)),
 		).
 		WithJobComponents(
-			utils.NewApplicationJobComponentBuilder().WithSchedulerPort(9999).WithName("calc-1").WithDockerfileName("calc.Dockerfile").WithSourceFolder("./calc/"),
-			utils.NewApplicationJobComponentBuilder().WithSchedulerPort(9999).WithName("calc-2").WithEnabled(true).WithDockerfileName("calc.Dockerfile").WithSourceFolder("./calc/"),
-			utils.NewApplicationJobComponentBuilder().WithSchedulerPort(9999).WithName("calc-3").WithEnabled(false).WithDockerfileName("calc.Dockerfile").WithSourceFolder("./calc/"),
-			utils.NewApplicationJobComponentBuilder().WithSchedulerPort(9999).WithName("calc-4").WithDockerfileName("calc.Dockerfile").WithSourceFolder("./calc2/"),
-			utils.NewApplicationJobComponentBuilder().WithSchedulerPort(9999).WithName("calc-5").WithEnabled(false).WithDockerfileName("calc.Dockerfile").WithSourceFolder("./calc2/"),
-			utils.NewApplicationJobComponentBuilder().WithSchedulerPort(9999).WithName("calc-6").WithEnabled(false).WithDockerfileName("calc.Dockerfile").WithSourceFolder("./calc3/").
+			utils.NewApplicationJobComponentBuilder().WithSchedulerPort(new(int32(9999))).WithName("calc-1").WithDockerfileName("calc.Dockerfile").WithSourceFolder("./calc/"),
+			utils.NewApplicationJobComponentBuilder().WithSchedulerPort(new(int32(9999))).WithName("calc-2").WithEnabled(true).WithDockerfileName("calc.Dockerfile").WithSourceFolder("./calc/"),
+			utils.NewApplicationJobComponentBuilder().WithSchedulerPort(new(int32(9999))).WithName("calc-3").WithEnabled(false).WithDockerfileName("calc.Dockerfile").WithSourceFolder("./calc/"),
+			utils.NewApplicationJobComponentBuilder().WithSchedulerPort(new(int32(9999))).WithName("calc-4").WithDockerfileName("calc.Dockerfile").WithSourceFolder("./calc2/"),
+			utils.NewApplicationJobComponentBuilder().WithSchedulerPort(new(int32(9999))).WithName("calc-5").WithEnabled(false).WithDockerfileName("calc.Dockerfile").WithSourceFolder("./calc2/"),
+			utils.NewApplicationJobComponentBuilder().WithSchedulerPort(new(int32(9999))).WithName("calc-6").WithEnabled(false).WithDockerfileName("calc.Dockerfile").WithSourceFolder("./calc3/").
 				WithEnvironmentConfig(utils.NewJobComponentEnvironmentBuilder().WithEnvironment(envName).WithEnabled(true)),
-			utils.NewApplicationJobComponentBuilder().WithSchedulerPort(9999).WithName("calc-7").WithEnabled(true).WithDockerfileName("calc.Dockerfile").WithSourceFolder("./calc4/").
+			utils.NewApplicationJobComponentBuilder().WithSchedulerPort(new(int32(9999))).WithName("calc-7").WithEnabled(true).WithDockerfileName("calc.Dockerfile").WithSourceFolder("./calc4/").
 				WithEnvironmentConfig(utils.NewJobComponentEnvironmentBuilder().WithEnvironment(envName).WithEnabled(false)),
-			utils.NewApplicationJobComponentBuilder().WithName("calc-8").WithSchedulerPort(9999).WithImage("calc-8-image"),
-			utils.NewApplicationJobComponentBuilder().WithName("calc-9").WithSchedulerPort(9999).WithImage("calc-9-image").
+			utils.NewApplicationJobComponentBuilder().WithName("calc-8").WithSchedulerPort(new(int32(9999))).WithImage("calc-8-image"),
+			utils.NewApplicationJobComponentBuilder().WithName("calc-9").WithSchedulerPort(new(int32(9999))).WithImage("calc-9-image").
 				WithEnvironmentConfig(utils.NewJobComponentEnvironmentBuilder().WithEnvironment(envName).WithEnabled(false)),
-			utils.NewApplicationJobComponentBuilder().WithName("calc-10").WithSchedulerPort(9999).WithImage("calc-10-image").WithEnabled(false).
+			utils.NewApplicationJobComponentBuilder().WithName("calc-10").WithSchedulerPort(new(int32(9999))).WithImage("calc-10-image").WithEnabled(false).
 				WithEnvironmentConfig(utils.NewJobComponentEnvironmentBuilder().WithEnvironment(envName).WithEnabled(true)),
 		).
 		BuildRA()
@@ -696,13 +696,13 @@ func (s *applyConfigTestSuite) Test_BuildAndDeployComponentImages_BuildChangedCo
 			utils.NewApplicationComponentBuilder().WithPort("any", 8080).WithName("comp-deployonly").WithImage("comp-deployonly:anytag"),
 		).
 		WithJobComponents(
-			utils.NewApplicationJobComponentBuilder().WithSchedulerPort(9999).WithName("job-changed").WithDockerfileName("job-changed.Dockerfile"),
-			utils.NewApplicationJobComponentBuilder().WithSchedulerPort(9999).WithName("job-new").WithDockerfileName("job-new.Dockerfile"),
-			utils.NewApplicationJobComponentBuilder().WithSchedulerPort(9999).WithName("job-unchanged").WithDockerfileName("job-unchanged.Dockerfile"),
-			utils.NewApplicationJobComponentBuilder().WithSchedulerPort(9999).WithName("job-common1-unchanged").WithDockerfileName("common1.Dockerfile"),
-			utils.NewApplicationJobComponentBuilder().WithSchedulerPort(9999).WithName("job-common2-changed").WithDockerfileName("common2.Dockerfile"),
-			utils.NewApplicationJobComponentBuilder().WithSchedulerPort(9999).WithName("job-common3-changed").WithDockerfileName("common3.Dockerfile"),
-			utils.NewApplicationJobComponentBuilder().WithSchedulerPort(9999).WithName("job-deployonly").WithImage("job-deployonly:anytag"),
+			utils.NewApplicationJobComponentBuilder().WithSchedulerPort(new(int32(9999))).WithName("job-changed").WithDockerfileName("job-changed.Dockerfile"),
+			utils.NewApplicationJobComponentBuilder().WithSchedulerPort(new(int32(9999))).WithName("job-new").WithDockerfileName("job-new.Dockerfile"),
+			utils.NewApplicationJobComponentBuilder().WithSchedulerPort(new(int32(9999))).WithName("job-unchanged").WithDockerfileName("job-unchanged.Dockerfile"),
+			utils.NewApplicationJobComponentBuilder().WithSchedulerPort(new(int32(9999))).WithName("job-common1-unchanged").WithDockerfileName("common1.Dockerfile"),
+			utils.NewApplicationJobComponentBuilder().WithSchedulerPort(new(int32(9999))).WithName("job-common2-changed").WithDockerfileName("common2.Dockerfile"),
+			utils.NewApplicationJobComponentBuilder().WithSchedulerPort(new(int32(9999))).WithName("job-common3-changed").WithDockerfileName("common3.Dockerfile"),
+			utils.NewApplicationJobComponentBuilder().WithSchedulerPort(new(int32(9999))).WithName("job-deployonly").WithImage("job-deployonly:anytag"),
 		).
 		BuildRA()
 	currentRd := utils.NewDeploymentBuilder().
@@ -825,7 +825,7 @@ func (s *applyConfigTestSuite) Test_BuildAndDeployComponentImages_DetectComponen
 			utils.NewApplicationComponentBuilder().WithPort("any", 8080).WithName("comp"),
 		).
 		WithJobComponents(
-			utils.NewApplicationJobComponentBuilder().WithSchedulerPort(9999).WithName("job"),
+			utils.NewApplicationJobComponentBuilder().WithSchedulerPort(new(int32(9999))).WithName("job"),
 		)
 	defaultRa := raBuilder.WithBuildSecrets(buildSecretName).BuildRA()
 	raWithoutSecret := raBuilder.WithBuildSecrets().BuildRA()
@@ -1454,7 +1454,7 @@ func (s *applyConfigTestSuite) Test_Deploy_JobWithMissingImageTagNameShouldFail(
 		WithAppName(appName).
 		WithEnvironment("dev", "anybranch").
 		WithJobComponents(
-			utils.NewApplicationJobComponentBuilder().WithSchedulerPort(9999).WithName("deployjob").WithImage("any:{imageTagName}"),
+			utils.NewApplicationJobComponentBuilder().WithSchedulerPort(new(int32(9999))).WithName("deployjob").WithImage("any:{imageTagName}"),
 		).
 		BuildRA()
 
@@ -1484,7 +1484,7 @@ func (s *applyConfigTestSuite) Test_Deploy_JobWithImageTagNameInRAShouldSucceed(
 		WithAppName(appName).
 		WithEnvironment("dev", "anybranch").
 		WithJobComponents(
-			utils.NewApplicationJobComponentBuilder().WithSchedulerPort(9999).WithName("deployjob").WithImage("any:{imageTagName}").
+			utils.NewApplicationJobComponentBuilder().WithSchedulerPort(new(int32(9999))).WithName("deployjob").WithImage("any:{imageTagName}").
 				WithEnvironmentConfig(utils.NewJobComponentEnvironmentBuilder().WithEnvironment("dev").WithImageTagName("anytag")),
 		).
 		BuildRA()
@@ -1512,7 +1512,7 @@ func (s *applyConfigTestSuite) Test_DeployComponentWitImageTagNameInPipelineArgS
 		WithAppName(appName).
 		WithEnvironment("dev", "anybranch").
 		WithJobComponents(
-			utils.NewApplicationJobComponentBuilder().WithSchedulerPort(9999).WithName("deployjob").WithImage("any:{imageTagName}"),
+			utils.NewApplicationJobComponentBuilder().WithSchedulerPort(new(int32(9999))).WithName("deployjob").WithImage("any:{imageTagName}"),
 		).
 		BuildRA()
 
@@ -1540,8 +1540,8 @@ func (s *applyConfigTestSuite) Test_Deploy_ComponentsToDeployValidation() {
 			utils.AnApplicationComponent().WithName("comp2").WithImage("some-image"),
 		).
 		WithJobComponents(
-			utils.AnApplicationJobComponent().WithName("job1").WithImage("some-image").WithSchedulerPort(8080),
-			utils.AnApplicationJobComponent().WithName("job2").WithImage("some-image").WithSchedulerPort(8080),
+			utils.AnApplicationJobComponent().WithName("job1").WithImage("some-image").WithSchedulerPort(new(int32(8080))),
+			utils.AnApplicationJobComponent().WithName("job2").WithImage("some-image").WithSchedulerPort(new(int32(8080))),
 		)
 	const appName = "anyapp"
 	rdBuilder := utils.NewDeploymentBuilder().WithAppName(appName).WithEnvironment("dev").
@@ -1549,8 +1549,8 @@ func (s *applyConfigTestSuite) Test_Deploy_ComponentsToDeployValidation() {
 			utils.NewDeployComponentBuilder().WithName("comp1").WithImage("some-image"),
 			utils.NewDeployComponentBuilder().WithName("comp2").WithImage("some-image")).
 		WithJobComponents(
-			utils.NewDeployJobComponentBuilder().WithName("job1").WithImage("some-image").WithSchedulerPort(8080),
-			utils.NewDeployJobComponentBuilder().WithName("job2").WithImage("some-image").WithSchedulerPort(8080))
+			utils.NewDeployJobComponentBuilder().WithName("job1").WithImage("some-image").WithSchedulerPort(new(int32(8080))),
+			utils.NewDeployJobComponentBuilder().WithName("job2").WithImage("some-image").WithSchedulerPort(new(int32(8080))))
 	activeRadixDeployment := rdBuilder.BuildRD()
 	ra := raBuilder.BuildRA()
 	rr := raBuilder.GetRegistrationBuilder().BuildRR()
@@ -1637,9 +1637,9 @@ func (s *applyConfigTestSuite) Test_DeployComponentImages_ImageTagNames() {
 				WithEnvironmentConfig(utils.NewComponentEnvironmentBuilder().WithEnvironment(envName).WithImageTagName("comp2envtag")),
 		).
 		WithJobComponents(
-			utils.NewApplicationJobComponentBuilder().WithSchedulerPort(9999).WithName("job1").WithImage("job1img:{imageTagName}").
+			utils.NewApplicationJobComponentBuilder().WithSchedulerPort(new(int32(9999))).WithName("job1").WithImage("job1img:{imageTagName}").
 				WithEnvironmentConfig(utils.NewJobComponentEnvironmentBuilder().WithEnvironment(envName).WithImageTagName("job1envtag")),
-			utils.NewApplicationJobComponentBuilder().WithSchedulerPort(9999).WithName("job2").WithImage("job2img:{imageTagName}").
+			utils.NewApplicationJobComponentBuilder().WithSchedulerPort(new(int32(9999))).WithName("job2").WithImage("job2img:{imageTagName}").
 				WithEnvironmentConfig(utils.NewJobComponentEnvironmentBuilder().WithEnvironment(envName).WithImageTagName("job2envtag")),
 		).
 		BuildRA()
@@ -1687,7 +1687,7 @@ func (s *applyConfigTestSuite) Test_Promote_DeployImagesSetFromSourceDeployment(
 			utils.NewApplicationComponentBuilder().WithPort("any", 8080).WithName("comp2").WithImage("any"),
 		).
 		WithJobComponents(
-			utils.NewApplicationJobComponentBuilder().WithSchedulerPort(9999).WithName("job1").WithImage("any"),
+			utils.NewApplicationJobComponentBuilder().WithSchedulerPort(new(int32(9999))).WithName("job1").WithImage("any"),
 		).
 		BuildRA()